Add shapely import guards to make sure tests run without extras
