+++ nucleus/metrics/cuboid_utils.py
@@ -6,27 +6,9 @@ import numpy as np
 try:
     from shapely.geometry import Polygon
 except ModuleNotFoundError:
-    import sys
+    from .shapely_not_installed import ShapelyNotInstalled
 
-    class Polygon:  # type: ignore
-        def __init__(self, *args, **kwargs):
-            """Object to make sure we only raise errors if actually trying to use shapely"""
-            if sys.platform.startswith("darwin"):
-                platform_specific_msg = (
-                    "Depending on Python environment used GEOS might need to be installed via "
-                    "`brew install geos`."
-                )
-            elif sys.platform.startswith("linux"):
-                platform_specific_msg = (
-                    "Depending on Python environment used GEOS might need to be installed via "
-                    "system package `libgeos-dev`."
-                )
-            else:
-                platform_specific_msg = "GEOS package will need to be installed see (https://trac.osgeo.org/geos/)"
-            raise ModuleNotFoundError(
-                f"Module 'shapely' not found. Install optionally with `scale-nucleus[shapely]` or when developing "
-                f"`poetry install -E shapely`. {platform_specific_msg}"
-            )
+    Polygon = ShapelyNotInstalled
 
 
 from nucleus.annotation import CuboidAnnotation
